requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
alembic
python-dotenv
python-json-logger
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
ctransformers
mlflow
peft
transformers
datasets
accelerate
bitsandbytes
trl
prometheus-client
